issue-11: mc1 now uses dummy field for cndx
Rather than round the concentration values given,
`mc1` now uses a dummy field ("rndc") to calculate
the concentration index. The rounding remains the
same, using 3 significant figures.

Multiple changes in mc3_methods, mc4, and tcplFit
will be needed to propogate the chagnes, because
most of the code is based on "conc" or "logc" --
not on "cndx."